feat: code commented and ordered
