+++ Erlang-Legacy/Assets/Core/UI/KeyIndicatorDisposer.cs.meta
