+++ Erlang-Legacy/Assets/Core/UI/KeyIndicatorTrigger.cs
@@ -1,0 +1,69 @@
+﻿using System.Collections;
+using Core.Shared.Enum;
+using UnityEngine;
+
+namespace Core.UI
+{
+    public class KeyIndicatorTrigger : MonoBehaviour
+    {
+        [SerializeField] GameKey gameKey;
+        [SerializeField] string function;
+        [SerializeField] float waitTimeBeforeShowingKeys;
+        private bool itsNeeded = true;
+        private bool playerIn = false;
+
+        //pre: --
+        //post: if player uses key in tutorial we make it disapear
+        private void Update()
+        {
+            if (playerIn && Input.GetKeyDown(gameKey.ToString().ToLower()))
+            {
+                itsNeeded = false;
+                playerIn = false; //we dont care anymore
+                KeyIndicatorDisposer.Instance?.HideTutorial();
+            }
+        }
+
+        //pre: 
+        //post: if player is in range and tutorial didn't had shown already
+        //      it start a coroutine that will show the instructions
+        private void OnTriggerEnter2D(Collider2D other)
+        {
+            if (!itsNeeded)
+                return;
+
+            if (other.tag == "Player" && !playerIn)
+            {
+                playerIn = true;
+                StartCoroutine(ShowKeyIndication());
+            }
+        }
+
+        //pre: 
+        //post: if player is out of range and tutorial it's beeing shown
+        //      it hides.
+        private void OnTriggerExit2D(Collider2D other)
+        {
+            if (!itsNeeded)
+                return;
+
+            if (other.tag == "Player" && playerIn)
+            {
+                KeyIndicatorDisposer.Instance?.HideTutorial();
+                playerIn = false;
+            }
+        }
+
+        //pre: --
+        //post: show tutorial instructions
+        IEnumerator ShowKeyIndication()
+        {
+            yield return new WaitForSeconds(waitTimeBeforeShowingKeys);
+
+            if (itsNeeded && playerIn)
+            {
+                KeyIndicatorDisposer.Instance?.ShowTutorial(gameKey, function);
+            }
+        }
+    }
+}
