+++ Erlang-Legacy/Assets/Core/UI/KeyIndicatorTrigger.cs.meta
