+++ Erlang-Legacy/Assets/Core/UI/KeyPressAnimation.cs.meta
